Move `ActionDashboardCard` type def to dashboard types file
